+++ apps/app/src/views/routes/PatientDetails/index.ts
@@ -1,0 +1,1 @@
+export * from './Patient';
